+++ src/main/java/edu/osu/waiting4ubackend/client/DBClient.java
@@ -2,6 +2,7 @@ package edu.osu.waiting4ubackend.client;
 
 import com.google.cloud.datastore.*;
 import edu.osu.waiting4ubackend.entity.Admin;
+import edu.osu.waiting4ubackend.entity.User;
 
 import java.io.IOException;
 
@@ -53,6 +54,70 @@ public class DBClient {
                 .setId(key.getId().toString())
                 .setUserName(adminEntity.getString("userName"))
                 .setEmail(adminEntity.getString("email"))
+                .build();
+    }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+    //user functions from here
+    private static final String USERS_COLLECTION_NAME = "users";
+
+    // functions for user from here
+    public boolean user_nameExists(String userName) {
+        Query<Entity> query =
+                Query.newEntityQueryBuilder().setKind(USERS_COLLECTION_NAME)
+                        .setFilter(StructuredQuery.PropertyFilter.eq("user_name", userName))
+                        .build();
+        QueryResults<Entity> results = db.run(query);
+        return results.hasNext();
+    }
+
+    public boolean user_emailExists(String email) {
+        Query<Entity> query =
+                Query.newEntityQueryBuilder().setKind(USERS_COLLECTION_NAME)
+                        .setFilter(StructuredQuery.PropertyFilter.eq("email", email))
+                        .build();
+        QueryResults<Entity> results = db.run(query);
+        return results.hasNext();
+    }
+
+    public String saveUser(User user) {
+        Key key = db.allocateId(db.newKeyFactory().setKind(USERS_COLLECTION_NAME).newKey());
+        Entity userEntity = Entity.newBuilder(key)
+                .set("user_name", user.getUserName())
+                .set("email", user.getEmail())
+                .set("password", user.getPassword())
+                .set("introduction", user.getIntroduction())
+                .build();
+        db.put(userEntity);
+
+        return key.getId().toString();
+    }
+
+    public User getUserById(long id) {
+        Key key = db.newKeyFactory().setKind(USERS_COLLECTION_NAME).newKey(id);
+        Entity userEntity = db.get(key);
+        if(userEntity == null) return null;
+        return new User.UserBuilder()
+                .setId(key.getId().toString())
+                .setUserName(userEntity.getString("user_name"))
+                .setEmail(userEntity.getString("email"))
+                .setIntroduction((userEntity.getString("introduction")))
                 .build();
     }
 }
